+++ doc/handle_request_example/android/app/src/main/kotlin/com/example/handle_request_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.handle_request_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}